+++ app/src/main/java/com/herdal/deezerapp/domain/repository/TrackRepository.kt
@@ -7,6 +7,6 @@ interface TrackRepository {
     suspend fun fetchTracksByAlbum(albumId: Int): List<Track>
     suspend fun addTrackToFavorite(track: Track)
     suspend fun deleteTrackFromFavorite(track: Track)
-    suspend fun isTrackFavorite(id: Int): Boolean
+    suspend fun isTrackFavorite(id: Long): Boolean
     fun getAllFavoriteTracks(): Flow<List<Track>>
 }
